Got search for a weapon's price working.
The issue with this is that it's only going to search once for each weapon in the string. Volume probably needs to be re-worked with a similar method, so that average prices can eventually be semi-accurate. I'm thinking of something that keeps looping through the listing string, consuming it until it can't find anything.